splitting useAuth to custom hooks
