+++ apps/client/src/components/PokeCard/PokeCard.tsx
@@ -3,14 +3,15 @@ import { Pokemon } from '../../types/pokemons';
 import { FavoriteButton } from './FavoriteButton';
 import { useUpdatePokemonCache } from '../../hooks/useUpdatePokemonCache';
 import { useFavoriteApiOperations } from '../../hooks/useFavoriteApiOperations';
-import { useAuth } from '../../hooks/auth/useAuth';
 import { isNotNullOrUndefined } from '../../utils';
 import toast from 'react-hot-toast';
+import { useSession } from '../../hooks/auth/useSession';
 
 export const PokeCard: React.FC<Pokemon> = React.memo(function PokeCard({ id, name, isFavorite }) {
-  const { session } = useAuth();
+  const { data: session, isLoading } = useSession();
   const updatePokemonCache = useUpdatePokemonCache();
   const { removeFavorite, addFavorite } = useFavoriteApiOperations();
+
   const onFavoriteClick = () => {
     if (isNotNullOrUndefined(session) && !session.loggedIn) {
       toast.error('Please login to favorite a pokemon');
