+++ apps/client/src/hooks/auth/useLogout.ts
@@ -1,0 +1,14 @@
+import { useMutation, UseMutationOptions } from '@tanstack/react-query';
+import AuthService from '../../requests/auth';
+import { useRefreshSession } from './useRefreshSession';
+
+export const useLogout = (options?: UseMutationOptions) => {
+  const refreshSession = useRefreshSession();
+
+  return useMutation({
+    mutationKey: ['register'],
+    mutationFn: AuthService.logout,
+    onSuccess: refreshSession,
+    ...options,
+  });
+};
