+++ apps/client/src/hooks/auth/useRefreshSession.ts
@@ -1,0 +1,10 @@
+import { useQueryClient } from '@tanstack/react-query';
+
+export const useRefreshSession = () => {
+  const queryClient = useQueryClient();
+
+  return async () => {
+    await queryClient.invalidateQueries({ queryKey: ['session'] });
+    await queryClient.invalidateQueries({ queryKey: ['pokemons'] });
+  };
+};
