+++ apps/client/src/types/auth.ts
@@ -1,0 +1,4 @@
+export type AuthResponse = {
+  message: string;
+  error?: string;
+};
